+++ viewer/src/components/Footer/FilePanel.jsx
@@ -5,14 +5,14 @@ import { faPrint } from '@fortawesome/free-solid-svg-icons';
 import Actions from '../../actions/actions';
 import { get } from '../../reducers';
 import FileBlock from '../FileBlock';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { ControlButton, TextButton } from '../StyledPrimitives';
 import { useTranslation } from '../Translation';
 import { useDispatch, useSelector } from 'react-redux';
 import ModalWindow from "../ModalWindows/ModalWindow";
 import { ActionTypes } from "../../constants";
 
-const style = css`
+const Root = styled.div`
     width: 100%;
     display: flex;
     justify-content: center;
@@ -37,42 +37,45 @@ export default () => {
     const t = useTranslation();
     const dispatch = useDispatch();
     const fileName = useSelector(get.fileName);
+    const isDocumentLoaded = useSelector(get.isDocumentLoaded);
     const [isNotificationShown, showNotification] = React.useState(false);
-    const { onSaveNotification } = useSelector(get.uiOptions);
+    const { onSaveNotification, hideOpenAndCloseButtons, hidePrintButton, hideSaveButton } = useSelector(get.uiOptions);
 
     const saveHandler = () => dispatch(Actions.tryToSaveDocument());
 
+    if (!isDocumentLoaded) return null;
+
     return (
-        <div css={style}>
-            {fileName ? (
+        <Root>
+            {hideOpenAndCloseButtons ? null :
                 <span title={t("Close document")}>
                     <ControlButton
                         onClick={() => dispatch(Actions.closeDocumentAction())}
                         icon={faTimesCircle}
                     />
-                </span>
-            ) : null}
-            <FileBlock fileName={fileName} />
-            {fileName ? (
-                <>
-                    <ControlButton
-                        icon={faPrint} title={t('Print document')}
-                        onClick={() => dispatch({ type: ActionTypes.OPEN_PRINT_DIALOG })}
-                    />
-                    <TextButton
-                        onClick={() => {
-                            if (onSaveNotification && onSaveNotification.text) {
-                                showNotification(true);
-                            } else {
-                                saveHandler();
-                            }
-                        }}
-                        title={t("Save document")}
-                    >
-                        {t('Save')}
-                    </TextButton>
-                </>
-            ) : null}
+                </span>}
+
+            {hideOpenAndCloseButtons ? fileName ? <span>{fileName}</span> : null :
+                <FileBlock fileName={fileName || ''} />}
+
+            {hidePrintButton ? null : <ControlButton
+                icon={faPrint} title={t('Print document')}
+                onClick={() => dispatch({ type: ActionTypes.OPEN_PRINT_DIALOG })}
+            />}
+
+            {hideSaveButton ? null : <TextButton
+                onClick={() => {
+                    if (onSaveNotification && onSaveNotification.text) {
+                        showNotification(true);
+                    } else {
+                        saveHandler();
+                    }
+                }}
+                title={t("Save document")}
+            >
+                {t('Save')}
+            </TextButton>}
+
             {isNotificationShown ?
                 <ModalWindow
                     onClose={() => {
@@ -102,6 +105,6 @@ export default () => {
                         </ButtonBlock>
                     </SaveNotification>
                 </ModalWindow> : null}
-        </div>
+        </Root>
     );
 }